et-ai-concierge/packages/agents/requirements.txt
fastapi
uvicorn[standard]
websockets
langgraph
langchain
langchain-groq
mem0ai
redis
psycopg2-binary
groq
edge-tts
finnhub-python
langsmith
sentry-sdk
pydantic
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
sentence-transformers
pgvector
